# RavenDB-4586