--- a/src/Voron/Data/EmptyIterator.cs
+++ b/src/Voron/Data/EmptyIterator.cs
@@ -32,7 +32,7 @@ namespace Voron.Data
         }
 
 
-        public event Action<IIterator> OnDispoal;
+        public event Action<IIterator> OnDisposal;
 
         public IEnumerable<string> DumpValues()
         {
@@ -67,9 +67,7 @@ namespace Voron.Data
 
         public void Dispose()
         {
-            var action = OnDispoal;
-            if (action != null)
-                action(this);
+            OnDisposal?.Invoke(this);
         }
     }
 }