--- a/test/FastTests/Voron/StorageTest.cs
+++ b/test/FastTests/Voron/StorageTest.cs
@@ -162,9 +162,8 @@ namespace FastTests.Voron
 
         protected unsafe Tuple<Slice, Slice> ReadKey(Transaction txh, Tree tree, Slice key)
         {
-            Lazy<TreeCursor> lazy;
             TreeNodeHeader* node;
-            var p = tree.FindPageFor(key, out node, out lazy);
+            var p = tree.FindPageFor(key, out node);
 
 
             if (node == null)