Avoid creating the lazy and all cursor objects when it is not needed.

In the cases where the cursor will be used, we now only allocate the lambda, we avoid allocating a Lazy on top of the lambda itself. The benchmarks show there are a few improvements in Tables.

```
Before:

fill rnd                   :      8.005 ms    124.906 ops / sec
fill rnd separate tx       :     28.719 ms     34.819 ops / sec
insert rnd separate tx     :      4.420 ms     22.622 ops / sec

After:

fill rnd                   :      5.909 ms    169.232 ops / sec
fill rnd separate tx       :     27.967 ms     35.755 ops / sec
insert rnd separate tx     :      4.055 ms     24.660 ops / sec
```